+++ cns/process/clustering.py
@@ -1,5 +1,5 @@
 import numpy as np
-from numba import jit
+from cns.utils.jit import jit
 from cns.process.segments import get_consecutive_segs
 from cns.utils.conversions import values_count
 from cns.utils.logging import log_info